Add initial scrape script
Currently, it is proxied trough a-watch.dont-break.it as a caching
reverse proxy. This vastly improved the speed of requests and as a
bonus we wont DOS the servers of abgeordnetenwatch.

We are currently not saving the parties or politicians, thats to-do.
Also, storing the constituancy and so on is also open.